Remove markdown blog pages and components and rename sanity post to blog
