+++ pages/blog.js
@@ -1,21 +1,31 @@
+import { useLiveQuery } from "next-sanity/preview";
+
+import Card from "components/Card";
+import { readToken } from "./api/sanity.api";
+import { getClient } from "./api/sanity.client";
+import { getPosts, postsQuery } from "./api/sanity.queries";
 import { Helmet, HeaderNav, Footer } from "@components";
-import { BlogPost } from "components/BlogPostPreview";
-import { BlogPostPreviewList } from "components/BlogPostPreviewList";
-import { useState, useEffect } from "react";
-import { getAllPosts } from "../lib/api";
+import { useRouter } from "next/router";
 
-export default function Blog(props) {
-  const [blogPosts, setBlogPosts] = useState([]);
+export const getStaticProps = async ({ draftMode = false }) => {
+  const client = getClient(draftMode ? { token: readToken } : undefined);
+  const posts = await getPosts(client);
 
-  useEffect(() => {
-    const blogPostObjects = props.allPosts;
-    const blogPosts = blogPostObjects.map(BlogPost.fromObject);
-    setBlogPosts(blogPosts);
-  }, [props]);
+  return {
+    props: {
+      draftMode,
+      token: draftMode ? readToken : "",
+      posts,
+    },
+  };
+};
 
+export default function Post(props) {
+  const router = useRouter();
+  const [posts] = useLiveQuery(props.posts, postsQuery);
   return (
     <div className="blog-page">
-      <Helmet />
+      <Helmet pageTitle={router.pathname} />
       <HeaderNav />
       <main>
         <header className="container hero">
@@ -23,30 +33,21 @@ export default function Blog(props) {
           <h1>Learn more about Software Systems at SFU</h1>
         </header>
         <section className="container">
-          <h3 className="category-title">Featured</h3>
-          <div className="posts-list featured">
-            <BlogPostPreviewList blogPosts={blogPosts} />
-          </div>
+          {posts.length ? (
+            <div>
+              <h3 className="category-title">Featured</h3>
+              <div className="posts-list featured">
+                {posts.map((post) => (
+                  <Card key={post._id} post={post} />
+                ))}
+              </div>
+            </div>
+          ) : (
+            <div>No post at the moment</div>
+          )}
         </section>
       </main>
       <Footer />
     </div>
   );
 }
-
-export const getStaticProps = async () => {
-  const allPosts = getAllPosts([
-    "title",
-    "date",
-    "slug",
-    "author",
-    "coverImage",
-    "excerpt",
-    "category",
-    "style",
-  ]);
-
-  return {
-    props: { allPosts },
-  };
-};
